recorder/requirements.txt
# AIMScribe Recorder Dependencies

# Audio recording
pyaudio>=0.2.13

# HTTP client
aiohttp>=3.9.0

# FastAPI server
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# System tray
pystray>=0.19.0
Pillow>=10.0.0
